Create verify hook
